Refactoring the documents tab related code and split components into seperate files
