+++ features/apimgt/org.wso2.carbon.apimgt.publisher.feature/src/main/resources/publisher/source/src/app/components/Apis/Details/index.js
@@ -22,7 +22,7 @@ import {Route, Switch, Redirect} from 'react-router-dom'
 import Overview from './Overview'
 import NavBar from './NavBar'
 import LifeCycle from './LifeCycle/LifeCycle'
-import Documents from './Documents'
+import Documents from './Documents/Documents'
 import {PageNotFound} from '../../Base/Errors/index'
 import Loading from '../../Base/Loading/Loading'
 import Resources from './Resources'
